[NUCLEO_F072RB] reorg hal folder
