+++ libraries/mbed/targets/hal/TARGET_STM/TARGET_STM32F0/PeripheralPins.h
@@ -38,6 +38,10 @@
 
 extern const PinMap PinMap_ADC[];
 
+//*** DAC ***
+
+extern const PinMap PinMap_DAC[];
+
 //*** I2C ***
 
 extern const PinMap PinMap_I2C_SDA[];
