+++ libraries/mbed/targets/hal/TARGET_STM/TARGET_STM32F0/gpio_object.h
@@ -64,6 +64,10 @@ static inline int gpio_read(gpio_t *obj)
     return ((*obj->reg_in & obj->mask) ? 1 : 0);
 }
 
+static inline int gpio_is_connected(const gpio_t *obj) {
+    return obj->pin != (PinName)NC;
+}
+
 #ifdef __cplusplus
 }
 #endif
