+++ libraries/mbed/targets/hal/TARGET_STM/TARGET_STM32F0/mbed_overrides.c
@@ -32,6 +32,9 @@ void mbed_sdk_init()
 {
     // Update the SystemCoreClock variable.
     SystemCoreClockUpdate();
+
+#if defined(TARGET_STM32F070RB) || defined(TARGET_STM32F072RB) || defined(TARGET_STM32F091RC)
     // Need to restart HAL driver after the RAM is initialized
     HAL_Init();
+#endif
 }
