+++ libraries/mbed/targets/hal/TARGET_STM/TARGET_STM32F0/sleep.c
@@ -33,6 +33,7 @@
 
 #include "cmsis.h"
 
+#if defined(TARGET_STM32F070RB)
 void sleep(void)
 {
     TIM_HandleTypeDef TimMasterHandle;
@@ -48,6 +49,23 @@ void sleep(void)
     // Enable HAL tick and us_ticker update interrupts
     __HAL_TIM_ENABLE_IT(&TimMasterHandle, (TIM_IT_CC2 | TIM_IT_UPDATE));
 }
+#else
+static TIM_HandleTypeDef TimMasterHandle;
+
+void sleep(void)
+{
+    TimMasterHandle.Instance = TIM2;
+
+    // Disable HAL tick interrupt
+    __HAL_TIM_DISABLE_IT(&TimMasterHandle, TIM_IT_CC2);
+
+    // Request to enter SLEEP mode
+    HAL_PWR_EnterSLEEPMode(PWR_MAINREGULATOR_ON, PWR_SLEEPENTRY_WFI);
+
+    // Enable HAL tick interrupt
+    __HAL_TIM_ENABLE_IT(&TimMasterHandle, TIM_IT_CC2);
+}
+#endif
 
 void deepsleep(void)
 {
